backend/requirements.txt
# Agent Orchestration
langgraph>=0.0.30
langchain>=0.1.0
langchain-openai>=0.0.2

# Database Drivers
psycopg2-binary>=2.9.9
neo4j>=5.16.0
qdrant-client>=1.7.0

# Object Storage
minio>=7.2.0

# API & Pydantic Validation
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-dotenv>=1.0.0

# Testing & Quality
pytest>=7.4.4
pytest-mock>=3.12.0
black>=24.1.1

# Additional Dependencies
httpx>=0.26.0
sqlalchemy>=2.0.25
pandas>=2.1.4